apps/order-service/requirements.txt
flask
pymongo
requests
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp
opentelemetry-instrumentation-requests
